Move entry-point main methods to separate package
