+++ src/main/java/com/turn/ttorrent/tracker/Tracker.java
@@ -17,30 +17,21 @@ package com.turn.ttorrent.tracker;
 
 import com.turn.ttorrent.common.Torrent;
 
-import java.io.File;
-import java.io.FilenameFilter;
 import java.io.IOException;
-import java.io.PrintStream;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.Collection;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
-import jargs.gnu.CmdLineParser;
-
-import org.apache.log4j.BasicConfigurator;
-import org.apache.log4j.ConsoleAppender;
-import org.apache.log4j.PatternLayout;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import org.simpleframework.transport.connect.Connection;
 import org.simpleframework.transport.connect.SocketConnection;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * BitTorrent tracker.
@@ -181,6 +172,13 @@ public class Tracker {
 	}
 
 	/**
+	 * Returns the list of tracker's torrents
+	 */
+	public Collection<TrackedTorrent> getTrackedTorrents() {
+		return torrents.values();
+	}
+
+	/**
 	 * Announce a new torrent on this tracker.
 	 *
 	 * <p>
@@ -316,84 +314,6 @@ public class Tracker {
 					// Ignore
 				}
 			}
-		}
-	}
-
-	/**
-	 * Display program usage on the given {@link PrintStream}.
-	 */
-	private static void usage(PrintStream s) {
-		s.println("usage: Tracker [options] [directory]");
-		s.println();
-		s.println("Available options:");
-		s.println("  -h,--help             Show this help and exit.");
-		s.println("  -p,--port PORT        Bind to port PORT.");
-		s.println();
-	}
-
-	/**
-	 * Main function to start a tracker.
-	 */
-	public static void main(String[] args) {
-		BasicConfigurator.configure(new ConsoleAppender(
-			new PatternLayout("%d [%-25t] %-5p: %m%n")));
-
-		CmdLineParser parser = new CmdLineParser();
-		CmdLineParser.Option help = parser.addBooleanOption('h', "help");
-		CmdLineParser.Option port = parser.addIntegerOption('p', "port");
-
-		try {
-			parser.parse(args);
-		} catch (CmdLineParser.OptionException oe) {
-			System.err.println(oe.getMessage());
-			usage(System.err);
-			System.exit(1);
-		}
-
-		// Display help and exit if requested
-		if (Boolean.TRUE.equals((Boolean)parser.getOptionValue(help))) {
-			usage(System.out);
-			System.exit(0);
-		}
-
-		Integer portValue = (Integer)parser.getOptionValue(port,
-			Integer.valueOf(DEFAULT_TRACKER_PORT));
-
-		String[] otherArgs = parser.getRemainingArgs();
-
-		if (otherArgs.length > 1) {
-			usage(System.err);
-			System.exit(1);
-		}
-
-		// Get directory from command-line argument or default to current
-		// directory
-		String directory = otherArgs.length > 0
-			? otherArgs[0]
-			: ".";
-
-		FilenameFilter filter = new FilenameFilter() {
-			@Override
-			public boolean accept(File dir, String name) {
-				return name.endsWith(".torrent");
-			}
-		};
-
-		try {
-			Tracker t = new Tracker(new InetSocketAddress(portValue.intValue()));
-
-			File parent = new File(directory);
-			for (File f : parent.listFiles(filter)) {
-				logger.info("Loading torrent from " + f.getName());
-				t.announce(TrackedTorrent.load(f));
-			}
-
-			logger.info("Starting tracker with {} announced torrents...",
-				t.torrents.size());
-			t.start();
-		} catch (Exception e) {
-			logger.error("{}", e.getMessage(), e);
-			System.exit(2);
 		}
 	}
 }
